space_cheer/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.13.5
cron-descriptor==1.4.5
cryptography==46.0.3
defusedxml==0.7.1
Django==5.1.7
django-allauth==65.13.1
django-celery-beat==2.7.0
django-crispy-forms==2.5
django-debug-toolbar==6.2.0
django-invitations==2.1.0
django-timezone-field==7.2.1
django-widget-tweaks==1.5.0
factory_boy==3.3.3
Faker==40.11.0
flower==2.0.1
gunicorn==25.1.0
humanize==4.15.0
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
packaging==26.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==5.1.0
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27; sys_platform != "win32"
python-magic-bin==0.4.14; sys_platform == "win32"
python3-openid==3.2.0
pytz==2026.1.post1
redis==5.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.4
tornado==6.5.5
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
